cleaning after discussion with Christian
